auto-format everything
This is long overdue and will become required if we move to any kind of
auto-linting/ci-linting.

I'm not completely sold on everything in here but I see no reason to
deviate from the designated standard. I'd need an effing good reason to
do that and I'd use this justification if this comes up in future
discussions.